修复我的关注书籍布局
