+++ core/infosphere-console/src/views/layouts/book/components/BookSidebar.vue
@@ -21,7 +21,6 @@
 import { defineComponent } from 'vue'
 import { cn } from '@/lib/utils.ts'
 import { BookIcon } from 'lucide-vue-next'
-import { useUserStore } from '@/stores/user.ts'
 
 export default defineComponent({
   name: 'BookSidebar',
@@ -45,17 +44,12 @@ export default defineComponent({
   methods: {
     initialize()
     {
-      const userStore = useUserStore()
-      const loggedIn = userStore.isLogin
       const items = [
         { title: '所有书籍', href: '/book/index' },
         { title: '公开书籍', href: '/book/public' },
-        { title: '私有书籍', href: '/book/private' }
+        { title: '私有书籍', href: '/book/private' },
+        { title: '我的关注', href: '/book/followed' }
       ]
-
-      if (loggedIn) {
-        items.push({ title: '我的关注', href: '/book/follow' })
-      }
       this.items = [...items] as any[]
     }
   }
